% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{s::can spectrolyzer output dataset import and processing}
\usage{
import(SerialNumber)
}
\arguments{
\item{SerialNumber}{Provide the serial number of the spectrolyzer wrapped in "".
This name is chosen because the default  output data file is name after the serial number of the spectroylyzer.}
}
\value{
A tibble of 6 columns with length equal the number of measurement the spectrolyzer performed
}
\description{
This function selects physically meaningful variables (Date/Time, turbidity, nitrate, toc, doc)
from the default output table of s::can spectrolyzers. This function will process them in a tidy way when
given the serial number of the spectroylyzer you want to import.Ensure the data file ".par" is present in
the Raw_data folder
}
\examples{
scan.Tools::import("09210014")

}
