% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{s::can spectrolyzer output dataset import and processing}
\usage{
import(path, serialnumber)
}
\arguments{
\item{path}{Provide the path to the ".par" dataset to be imported, wrapped in "".
If your dataset is still in ".fp" format, you need to visualize the fingerprint
through the Analyte software to be converted to ".par" format which is the parameter
file}

\item{serialnumber}{Provide the serial number of the spectrolyzer wrapped in "". This will be used to generate
a new column, named SN, populated with the provided serial number to follow the dataset}
}
\value{
A tibble of 6 columns with length equal the number of measurement the spectrolyzer performed
}
\description{
This function selects physically meaningful variables (Date/Time, turbidity, nitrate, toc, doc)
from the default output table of s::can spectrolyzers. This function will process them in a tidy way when
given the serial number of the spectroylyzer you want to import.Ensure the data file ".par" is present in
the Raw_data folder
}
\examples{
import("https://raw.githubusercontent.com/BrianY-Wang/foo/main/Raw_data/09210160.par", "09210160")

}
